Fixed some compiler warnings identified by gcc's -Wextra option.
The following issues have been addressed:

 * comparison between signed and unsigned - this was found in several places
   throughout the code and has been fixed in various ways
 * missing initializer - an incomplete initializer has been used for two
   struct instances in perl.c
 * unused parameter - when applicable, the parameter has been removed; in
   thirteen cases the parameter is required by different library API's and in
   two cases the parameter was left in place to retain a consistent interface
   within the affected modules; as __attribute__((unused)) is a GNU extension,
   it has not been used to document those exceptions

+++ src/cpufreq.c
@@ -40,7 +40,7 @@ static int cpufreq_init (void)
 		status = snprintf (filename, sizeof (filename),
 				"/sys/devices/system/cpu/cpu%d/cpufreq/"
 				"scaling_cur_freq", num_cpu);
-    		if (status < 1 || status >= sizeof (filename))
+		if ((status < 1) || ((unsigned int)status >= sizeof (filename)))
 			break;
 
 		if (access (filename, R_OK))
@@ -90,7 +90,7 @@ static int cpufreq_read (void)
 		status = snprintf (filename, sizeof (filename),
 				"/sys/devices/system/cpu/cpu%d/cpufreq/"
 				"scaling_cur_freq", i);
-    		if (status < 1 || status >= sizeof (filename))
+		if ((status < 1) || ((unsigned int)status >= sizeof (filename)))
 			return (-1);
 
 		if ((fp = fopen (filename, "r")) == NULL)
